fermeture possible une seule fois

+++ disqueAlberti.py
@@ -18,6 +18,7 @@ cmd = 0x40  # control byte to enable analog output
 #cryptogram
 cryptogram = "QbinxmFbxudssigyyutscooNcudty"
 pos = 0
+run = True
 
 # LCD
 mylcd = I2C_LCD_driver.lcd()
@@ -52,7 +53,7 @@ def buttonsSetup():
     GPIO.setup(shutdownButton, GPIO.IN, pull_up_down = GPIO.PUD_DOWN)
     GPIO.add_event_detect(shutdownButton, GPIO.RISING,
                           callback = shutdown,
-                          bouncetime = 200)
+                          bouncetime = 2000)
 
 def onMessage(client, userdata, message):
     global cryptogram
@@ -118,7 +119,7 @@ def loop():
         value = analogRead(0)
         analogWrite(value)
         newKey = int((256 - value) / 255 * alberti.Disk.size - 1)
-        if key != newKey or refresh :
+        if (key != newKey or refresh) and run:
             os.system('clear')
             refresh = False
             key = newKey
@@ -138,44 +139,47 @@ def get_ip():
     return subprocess.check_output(cmd, shell=True).decode("utf-8").strip()
 
 def shutdown(channel):
+    global run
+    if run :
+        run = False
+        try :
+            mqttClient.loop_stop()
+            imprimer('Fermeture', 'Debrancher apres')
+            time.sleep(2)
+            imprimer('Debrancher apres', '30 secondes')
+            time.sleep(2)
+            mylcd.lcd_clear()
+            mylcd.backlight(0)
+        except Exception as e :
+            print(str(e))
+        os.system("sudo shutdown -h now")
+
+def printConfiguration() :
+    imprimer('Demarrage', '')
+    time.sleep(1)
+    imprimer(datetime.datetime.now().strftime('%d %b %H:%M:%S'), get_ip())
+    time.sleep(3)
+    imprimer('MQTT BROKER ' + str(mqttPort), mqttBroker)
+    time.sleep(2)
+    imprimer('Message Topic', mqttMessageTopic)
+    time.sleep(2)
+    imprimer('Crypto Topic', mqttCryptoTopic)
+    time.sleep(2)
+
+def setup() :
     try :
-        imprimer('Shutting Down', '')
-        mqttClient.loop_stop()
-        time.sleep(1)
-        mylcd.lcd_clear()
-        mylcd.backlight(0)
-        time.sleep(1)
+        initMQTT()
+        imprimer('Connexion MQTT', 'reussie')
     except Exception as e :
-        print(str(e))
-    os.system("sudo shutdown -h now")
-  
-### START ###
-pos = 0
+        print(e)
+        imprimer('Connexion MQTT', 'echouee')
+    time.sleep(2)
 
-# Print configuration
-imprimer('Demarrage', '')
-time.sleep(1)
-imprimer('MQTT BROKER ' + str(mqttPort), mqttBroker)
-time.sleep(2)
-imprimer('Message Topic', mqttMessageTopic)
-time.sleep(1)
-imprimer('Crypto Topic', mqttCryptoTopic)
-time.sleep(1)
-
-try :
-    initMQTT()
-    imprimer('Connexion MQTT', 'reussie')
-except Exception as e :
-    print(e)
-    imprimer('Connexion MQTT', 'echouee')
-time.sleep(1)
-
-imprimer(datetime.datetime.now().strftime('%d %b %H:%M:%S'), get_ip())
-time.sleep(5)
-
-buttonsSetup()
+    buttonsSetup()
 
 # Start
+printConfiguration()
+setup()
 if len(sys.argv) > 1 :
     cryptogram = alberti.cleanupCryptogram(sys.argv[1])
